치유의 기록 업데이트
-제공받은 구름 정보 업데이트
-Dialog 업데이트
