+++ Cloud_Factory/Assets/Data/SeasonDate.json
@@ -1,5 +1,5 @@
 {
-    "mSecond": 16.067323684692384,
+    "mSecond": 2.1823198795318605,
     "mDay": 1,
     "mWeek": 0,
     "mSeason": 1,
