+++ Cloud_Factory/Assets/Scripts/KCH/object/EmptyObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ea681fd50a3a4fb4995a89ed953f3f3d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
