--- Cloud_Factory/Assets/Scripts/LJH/Record Of Healing/ProfileManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2dfed7cfc71f4494ea074b4818b97fcc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
