+++ Cloud_Factory/Library/StateCache/SceneView/8c/8cd7c613bf844de3b80696e27a479d5e.json
@@ -1,1 +1,1 @@
-﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":479.9307556152344,"y":351.69091796875,"z":-75.27214050292969},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":463.63751220703127,"orthographic":true}
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":true,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":true,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":342.48193359375,"y":277.28424072265627,"z":-217.43002319335938},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":290.89801025390627,"orthographic":true}
