+++ Cloud_Factory/Library/StateCache/SceneView/b1/b1a9c72ea2783f125661f1df1dc562b2.json
@@ -1,0 +1,1 @@
+﻿{"cameraMode":{"drawMode":0,"name":"Shaded","section":"Shading Mode"},"sceneLighting":false,"audioPlay":false,"sceneViewState":{"m_AlwaysRefresh":false,"showFog":true,"showSkybox":false,"showFlares":true,"showImageEffects":true,"showParticleSystems":true,"showVisualEffectGraphs":true,"m_FxEnabled":true},"in2DMode":true,"pivot":{"x":952.3804931640625,"y":148.5,"z":0.0},"rotation":{"x":0.0,"y":0.0,"z":0.0,"w":1.0},"size":70.71067810058594,"orthographic":true}
