+++ Cloud_Factory/UserSettings/EditorUserSettings.asset
@@ -15,22 +15,22 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c7c1c19021c6a0a292524223b047df7ee3d2cfb
       flags: 0
     RecentlyUsedScenePath-3:
-      value: 22424703114646680e0b0227036c6d1515180a2e6c072070052c1c3febee3376f7e93ffdfe
-      flags: 0
-    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c7b02170011242b68143f22245326ece92021
       flags: 0
-    RecentlyUsedScenePath-5:
+    RecentlyUsedScenePath-4:
       value: 22424703114646680e0b0227036c7c1c19021c6a1f3c29222c2e187df7ee3d2cfb
       flags: 0
-    RecentlyUsedScenePath-6:
+    RecentlyUsedScenePath-5:
       value: 22424703114646680e0b0227036c7b1515183b26233d227e38271427fb
       flags: 0
-    RecentlyUsedScenePath-7:
+    RecentlyUsedScenePath-6:
       value: 22424703114646680e0b0227036c78190012580920273334633c133af6f9
       flags: 0
-    RecentlyUsedScenePath-8:
+    RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c6c0017141d6a032e66072828093be7f27a2decee22f0
+      flags: 0
+    RecentlyUsedScenePath-8:
+      value: 22424703114646680e0b0227036c6d1515180a2e6c072070052c1c3febee3376f7e93ffdfe
       flags: 0
     RecentlyUsedScenePath-9:
       value: 22424703114646680e0b0227036c731f1415016439262f2434
